fix(Core/Unit): dodges, parries and critters not increasing weapon skills (#5895)
This fixes not getting skill ups from attacking an enemy and them parrying your attack or dodging it.
This also fixes attacking a critter and not getting weapon skill ups.
Previously Dodge and Parry were only giving a chance to increase the victims DEFENSE skill.
I changed it so it will proc on victim and attacker. for weapon and defense skills
Also, critters were being checked to not give weapon skill ups on purpose but I removed the check.