fix(DB/gameobject): Improve Draenei Vessel (#14339)
Asking for a test because my DB is quite polluted these days and I may
have forgot to set a value somewhere

<!-- First of all, THANK YOU for your contribution. -->

## Changes Proposed:
- Sniffs spawns for Draenei Vessel (already sniffed before)
- Pools normal Vessel them with trapped Vessel at 50% chance (confirmed
in sniff)
- Prevent trap from triggering at distance
- Fixes trapped Vessel being unlootable
- Fixes campfire with bad z coords using sniffed data

## Issues Addressed:
<!-- If your fix has a relating issue, link it below -->
- Closes https://github.com/azerothcore/azerothcore-wotlk/issues/14332
- Closes https://github.com/chromiecraft/chromiecraft/issues/4618

## SOURCE:
<!-- If you can, include a source that can strengthen your claim -->
WotLK Classic

## Tests Performed:
<!-- Does it build without errors? Did you test in-game? What did you
test? On which OS did you test? Describe any other tests performed -->
- Tested ingame
- 


## How to Test the Changes:
<!-- Describe in a detailed step-by-step order how to test the changes
-->

1. `.quest add 10028`
2. `.tele Tuurem`
3. see if behaviour improves

## Known Issues and TODO List:
<!-- Is there anything else left to do after this PR? -->

- [ ] I haven't found a way to make a truly empty loot table without
worldserver errors so I duplicated the table for normal vessel and
lowered the chance to 0.1 (I've seen no loot in WotLK Classic, but it
was lootable)
- [ ]

<!-- If you intend to contribute repeatedly to our project, it is a good
idea to join our discord channel. We set ranks for our contributors and
give them access to special resources or knowledge:
https://discord.com/invite/DasJqPba)
Do not remove the instructions below about testing, they will help users
to test your PR -->
## How to Test AzerothCore PRs
 
When a PR is ready to be tested, it will be marked as **[WAITING TO BE
TESTED]**.

You can help by testing PRs and writing your feedback here on the PR's
page on GitHub. Follow the instructions here:

http://www.azerothcore.org/wiki/How-to-test-a-PR

**REMEMBER**: when testing a PR that changes something **generic** (i.e.
a part of code that handles more than one specific thing), the tester
should not only check that the PR does its job (e.g. fixing spell XXX)
but **especially** check that the PR does not cause any regression (i.e.
introducing new bugs).

**For example**: if a PR fixes spell X by changing a part of code that
handles spells X, Y, and Z, we should not only test X, but **we should
test Y and Z as well**.