fix(Docker): config loading for ac-dev-server container (#14755)
## Changes Proposed:
-  This properly moves things from `DOCKER_ETC_FOLDER` to `confDir`
- originally broken with
https://github.com/azerothcore/azerothcore-wotlk/commit/4f4e4263509fae890a241e3ea9dd7fc9d10d831e

## Issues Addressed:
- Should hopefully close
https://github.com/azerothcore/acore-docker/issues/12

## SOURCE:

## Tests Performed:
- I'm not exactly aware of the testing procedure for this, as I don't
use acore-docker
- That said, the previous behavior has been replicated:

https://github.com/azerothcore/azerothcore-wotlk/blob/163ecc7c55549212739688b21941e37c9ecbb3f8/apps/compiler/includes/functions.sh#L148-L150
- Open to suggestions


## How to Test the Changes:
1. Open to Suggestions

## Known Issues and TODO List:

- Test properly

## How to Test AzerothCore PRs
 
When a PR is ready to be tested, it will be marked as **[WAITING TO BE
TESTED]**.

You can help by testing PRs and writing your feedback here on the PR's
page on GitHub. Follow the instructions here:

http://www.azerothcore.org/wiki/How-to-test-a-PR

**REMEMBER**: when testing a PR that changes something **generic** (i.e.
a part of code that handles more than one specific thing), the tester
should not only check that the PR does its job (e.g. fixing spell XXX)
but **especially** check that the PR does not cause any regression (i.e.
introducing new bugs).

**For example**: if a PR fixes spell X by changing a part of code that
handles spells X, Y, and Z, we should not only test X, but **we should
test Y and Z as well**.